+++ src/main/java/nl/underkoen/jetbrains/adventofcode/file/AOCFileEditor.java
@@ -1,0 +1,92 @@
+package nl.underkoen.jetbrains.adventofcode.file;
+
+import com.intellij.openapi.fileEditor.FileEditor;
+import com.intellij.openapi.fileEditor.FileEditorLocation;
+import com.intellij.openapi.fileEditor.FileEditorState;
+import com.intellij.openapi.util.Key;
+import com.intellij.openapi.vfs.VirtualFile;
+import com.intellij.ui.jcef.JBCefBrowser;
+import lombok.RequiredArgsConstructor;
+import nl.underkoen.jetbrains.adventofcode.general.Day;
+import org.jetbrains.annotations.Nls;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import javax.swing.*;
+import java.beans.PropertyChangeListener;
+
+public class AOCFileEditor implements FileEditor {
+    private final Day file;
+    private final JBCefBrowser browser;
+
+    public AOCFileEditor(Day file) {
+        this.file = file;
+        browser = new JBCefBrowser(String.format("https://adventofcode.com/%s/day/%s", file.getYear().getYear(), file.getDay()));
+    }
+
+    @Override
+    public @NotNull JComponent getComponent() {
+//        browser.setZoomLevel(0.7);
+        return browser.getComponent();
+    }
+
+    @Override
+    public @Nullable JComponent getPreferredFocusedComponent() {
+        return null;
+    }
+
+    @Override
+    public @Nls(capitalization = Nls.Capitalization.Title) @NotNull String getName() {
+        return "test";
+    }
+
+    @Override
+    public void setState(@NotNull FileEditorState state) {
+
+    }
+
+    @Override
+    public boolean isModified() {
+        return false;
+    }
+
+    @Override
+    public boolean isValid() {
+        return true;
+    }
+
+    @Override
+    public void addPropertyChangeListener(@NotNull PropertyChangeListener listener) {
+
+    }
+
+    @Override
+    public void removePropertyChangeListener(@NotNull PropertyChangeListener listener) {
+
+    }
+
+    @Override
+    public @Nullable FileEditorLocation getCurrentLocation() {
+        return null;
+    }
+
+    @Override
+    public void dispose() {
+
+    }
+
+    @Override
+    public <T> @Nullable T getUserData(@NotNull Key<T> key) {
+        return null;
+    }
+
+    @Override
+    public <T> void putUserData(@NotNull Key<T> key, @Nullable T value) {
+
+    }
+
+    @Override
+    public @Nullable VirtualFile getFile() {
+        return file;
+    }
+}