+++ src/main/java/nl/underkoen/jetbrains/adventofcode/file/AOCFileEditorProvider.java
@@ -1,0 +1,33 @@
+package nl.underkoen.jetbrains.adventofcode.file;
+
+import com.intellij.openapi.fileEditor.FileEditor;
+import com.intellij.openapi.fileEditor.FileEditorPolicy;
+import com.intellij.openapi.fileEditor.FileEditorProvider;
+import com.intellij.openapi.project.DumbAware;
+import com.intellij.openapi.project.Project;
+import com.intellij.openapi.vfs.VirtualFile;
+import nl.underkoen.jetbrains.adventofcode.general.Day;
+import org.jetbrains.annotations.NonNls;
+import org.jetbrains.annotations.NotNull;
+
+public class AOCFileEditorProvider implements FileEditorProvider, DumbAware {
+    @Override
+    public boolean accept(@NotNull Project project, @NotNull VirtualFile file) {
+        return file instanceof Day;
+    }
+
+    @Override
+    public @NotNull FileEditor createEditor(@NotNull Project project, @NotNull VirtualFile file) {
+        return new AOCFileEditor((Day) file);
+    }
+
+    @Override
+    public @NotNull @NonNls String getEditorTypeId() {
+        return "test";
+    }
+
+    @Override
+    public @NotNull FileEditorPolicy getPolicy() {
+        return FileEditorPolicy.HIDE_DEFAULT_EDITOR;
+    }
+}