+++ src/main/java/nl/underkoen/jetbrains/adventofcode/general/AOCTreeModal.java
@@ -1,0 +1,129 @@
+package nl.underkoen.jetbrains.adventofcode.general;
+
+import com.intellij.openapi.vfs.VirtualFile;
+import com.intellij.openapi.vfs.VirtualFileListener;
+import com.intellij.openapi.vfs.VirtualFileSystem;
+import org.jetbrains.annotations.NonNls;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import javax.swing.event.TreeModelListener;
+import javax.swing.tree.MutableTreeNode;
+import javax.swing.tree.TreeModel;
+import javax.swing.tree.TreeNode;
+import javax.swing.tree.TreePath;
+import java.io.IOException;
+
+public class AOCTreeModal extends VirtualFileSystem implements TreeModel {
+    private final AdventOfCode adventOfCode;
+
+    public AOCTreeModal() {
+        this.adventOfCode = new AdventOfCode(this);
+    }
+
+    @Override
+    public Object getRoot() {
+        return adventOfCode;
+    }
+
+    @Override
+    public Object getChild(Object parent, int index) {
+        return ((TreeNode)parent).getChildAt(index);
+    }
+
+    @Override
+    public int getChildCount(Object parent) {
+        return ((TreeNode)parent).getChildCount();
+    }
+
+    @Override
+    public boolean isLeaf(Object node) {
+        return ((TreeNode)node).isLeaf();
+    }
+
+    @Override
+    public void valueForPathChanged(TreePath path, Object newValue) {
+    }
+
+    @Override
+    public int getIndexOfChild(Object parent, Object child) {
+        if(parent == null || child == null)
+            return -1;
+        return ((TreeNode)parent).getIndex((TreeNode)child);
+    }
+
+    @Override
+    public void addTreeModelListener(TreeModelListener l) {
+
+    }
+
+    @Override
+    public void removeTreeModelListener(TreeModelListener l) {
+
+    }
+
+    @Override
+    public @NonNls @NotNull String getProtocol() {
+        return "aoc";
+    }
+
+    @Override
+    public @Nullable VirtualFile findFileByPath(@NotNull @NonNls String path) {
+        return null;
+    }
+
+    @Override
+    public void refresh(boolean asynchronous) {
+
+    }
+
+    @Override
+    public @Nullable VirtualFile refreshAndFindFileByPath(@NotNull String path) {
+        return null;
+    }
+
+    @Override
+    public void addVirtualFileListener(@NotNull VirtualFileListener listener) {
+
+    }
+
+    @Override
+    public void removeVirtualFileListener(@NotNull VirtualFileListener listener) {
+
+    }
+
+    @Override
+    protected void deleteFile(Object requestor, @NotNull VirtualFile vFile) throws IOException {
+        throw new IOException();
+    }
+
+    @Override
+    protected void moveFile(Object requestor, @NotNull VirtualFile vFile, @NotNull VirtualFile newParent) throws IOException {
+        throw new IOException();
+    }
+
+    @Override
+    protected void renameFile(Object requestor, @NotNull VirtualFile vFile, @NotNull String newName) throws IOException {
+        throw new IOException();
+    }
+
+    @Override
+    protected @NotNull VirtualFile createChildFile(Object requestor, @NotNull VirtualFile vDir, @NotNull String fileName) throws IOException {
+        throw new IOException();
+    }
+
+    @Override
+    protected @NotNull VirtualFile createChildDirectory(Object requestor, @NotNull VirtualFile vDir, @NotNull String dirName) throws IOException {
+        throw new IOException();
+    }
+
+    @Override
+    protected @NotNull VirtualFile copyFile(Object requestor, @NotNull VirtualFile virtualFile, @NotNull VirtualFile newParent, @NotNull String copyName) throws IOException {
+        throw new IOException();
+    }
+
+    @Override
+    public boolean isReadOnly() {
+        return true;
+    }
+}