+++ src/main/java/nl/underkoen/jetbrains/adventofcode/general/AdventOfCode.java
@@ -1,16 +1,27 @@
 package nl.underkoen.jetbrains.adventofcode.general;
 
-import nl.underkoen.jetbrains.adventofcode.general.tree.AbstractTreeNode;
+import com.intellij.openapi.util.NlsSafe;
+import com.intellij.openapi.vfs.VirtualFileSystem;
+import lombok.Getter;
+import nl.underkoen.jetbrains.adventofcode.file.VirtualDirectory;
+import org.jetbrains.annotations.NotNull;
 
 import javax.swing.tree.TreeNode;
 import java.time.LocalDateTime;
 import java.time.Month;
 import java.time.ZoneId;
+import java.util.Enumeration;
 import java.util.Vector;
 
-public class AdventOfCode extends AbstractTreeNode<Year, TreeNode> {
-    public AdventOfCode() {
-        super(null);
+public class AdventOfCode extends VirtualDirectory implements TreeNode {
+    @Getter private AOCTreeModal root;
+    private final Vector<Year> years;
+
+    public AdventOfCode(AOCTreeModal root) {
+        super(null, null);
+        this.root = root;
+        this.years = construct();
+        setFiles(years);
     }
 
     protected Vector<Year> construct() {
@@ -26,8 +37,57 @@ public class AdventOfCode extends AbstractTreeNode<Year, TreeNode> {
         return years;
     }
 
+    public Year getYear(int year) {
+        return years.get(year - 2015);
+    }
+
+    @Override
+    public TreeNode getChildAt(int childIndex) {
+        return years.get(childIndex);
+    }
+
+    @Override
+    public int getChildCount() {
+        return years.size();
+    }
+
+    @Override
+    public @NotNull @NlsSafe String getName() {
+        return "AdventOfCode";
+    }
+
+    @Override
+    public @NotNull VirtualFileSystem getFileSystem() {
+        return root;
+    }
+
+    @Override
+    public AdventOfCode getParent() {
+        return null;
+    }
+
+    @Override
+    public int getIndex(TreeNode node) {
+        return years.indexOf(node);
+    }
+
+    @Override
+    public boolean getAllowsChildren() {
+        return true;
+    }
+
+    @Override
+    public boolean isLeaf() {
+        return false;
+    }
+
+    @Override
+    public Enumeration<? extends TreeNode> children() {
+        return years.elements();
+    }
+
     @Override
     public String toString() {
-        return "AdventOfCode";
+        return getName();
     }
 }