+++ src/main/java/nl/underkoen/jetbrains/adventofcode/general/components/AdventOfCodeTreeCellRenderer.java
@@ -1,7 +1,10 @@
-package nl.underkoen.jetbrains.adventofcode.general;
+package nl.underkoen.jetbrains.adventofcode.general.components;
 
 import com.intellij.ui.JBColor;
-import com.intellij.util.ui.JBUI;
+import com.intellij.ui.treeStructure.Tree;
+import nl.underkoen.jetbrains.adventofcode.general.Day;
+import nl.underkoen.jetbrains.adventofcode.general.Year;
+import nl.underkoen.jetbrains.adventofcode.general.components.DayPanel;
 
 import javax.swing.*;
 import javax.swing.tree.TreeCellRenderer;
@@ -10,7 +13,7 @@ import java.awt.*;
 public class AdventOfCodeTreeCellRenderer implements TreeCellRenderer {
     private final TreeCellRenderer cellRenderer;
 
-    public AdventOfCodeTreeCellRenderer(JTree tree) {
+    public AdventOfCodeTreeCellRenderer(Tree tree) {
         this.cellRenderer = tree.getCellRenderer();
         tree.setCellRenderer(this);
     }
@@ -18,16 +21,7 @@ public class AdventOfCodeTreeCellRenderer implements TreeCellRenderer {
     @Override
     public Component getTreeCellRendererComponent(JTree tree, Object value, boolean selected, boolean expanded, boolean leaf, int row, boolean hasFocus) {
         if (value instanceof Day) {
-            Day day = (Day) value;
-
-            JPanel panel = new JPanel(new FlowLayout(FlowLayout.LEFT));
-            JLabel stars = new JLabel(day.starsString());
-            stars.setForeground(JBColor.yellow);
-            stars.setFont(stars.getFont().deriveFont(Font.BOLD));
-
-            panel.add(new JLabel(day.toString()));
-            panel.add(stars);
-            return panel;
+            return new DayPanel((Day) value);
         } else if (value instanceof Year) {
             Year year = (Year) value;
 
@@ -42,6 +36,7 @@ public class AdventOfCodeTreeCellRenderer implements TreeCellRenderer {
 
             panel.add(text);
             panel.add(stars);
+
             return panel;
         }
         return cellRenderer.getTreeCellRendererComponent(tree, value, selected, expanded, leaf, row, hasFocus);