+++ src/main/java/nl/underkoen/jetbrains/adventofcode/general/components/Icons.java
@@ -1,0 +1,11 @@
+package nl.underkoen.jetbrains.adventofcode.general.components;
+
+import com.intellij.openapi.util.IconLoader;
+
+import javax.swing.*;
+
+public class Icons {
+    private Icons() {};
+
+    public static final Icon AOC_FILE = IconLoader.getIcon("/icons/aoc.png", Icons.class);
+}