+++ src/main/java/nl/underkoen/jetbrains/adventofcode/toolwindows/AdventOfCodeViewerFactory.java
@@ -10,9 +10,9 @@ import org.jetbrains.annotations.NotNull;
 public class AdventOfCodeViewerFactory implements ToolWindowFactory {
     @Override
     public void createToolWindowContent(@NotNull Project project, @NotNull ToolWindow toolWindow) {
-        AdventOfCodeViewer viewer = new AdventOfCodeViewer(toolWindow);
+        AdventOfCodeViewer viewer = new AdventOfCodeViewer(project, toolWindow);
         ContentFactory contentFactory = ContentFactory.SERVICE.getInstance();
-        Content content = contentFactory.createContent(viewer.getContent(),"Challenges", false);
+        Content content = contentFactory.createContent(viewer.getContent(),null, false);
         toolWindow.getContentManager().addContent(content);
     }
 }